composer seems alright, for now exposing all ports for easier testing until there is no client

+++ server/src/main/java/com/codecool/sv_server/controller/AuthController.java
@@ -22,7 +22,7 @@ public class UserController {
     public ResponseEntity<SignupResponseDto> signup(
             @RequestBody SignupRequestDto signupRequestDto) {
         System.out.println(signupRequestDto.email() + " " + signupRequestDto.password());
-        int id = userService.signup(signupRequestDto);
+        long id = userService.signup(signupRequestDto);
         return ResponseEntity.ok(new SignupResponseDto(signupRequestDto.email(), id));
     }
 }
